+++ disco/voice/__init__.py
@@ -1,2 +1,3 @@
 from disco.voice.client import *
 from disco.voice.player import *
+from disco.voice.playable import *
