+++ disco/voice/opus.py
@@ -156,125 +156,3 @@ class OpusEncoder(BaseOpus):
 
 class OpusDecoder(BaseOpus):
     pass
-
-
-class BufferedOpusEncoder(OpusEncoder):
-    def __init__(self, source, *args, **kwargs):
-        self.source = source
-        self.frames = Queue(kwargs.pop('queue_size', 4096))
-        super(BufferedOpusEncoder, self).__init__(*args, **kwargs)
-        gevent.spawn(self._encoder_loop)
-
-    def _encoder_loop(self):
-        while self.source:
-            raw = self.source.read(self.frame_size)
-            if len(raw) < self.frame_size:
-                break
-
-            self.frames.put(self.encode(raw, self.samples_per_frame))
-            gevent.idle()
-        self.source = None
-
-    def have_frame(self):
-        return self.source or not self.frames.empty()
-
-    def next_frame(self):
-        return self.frames.get()
-
-
-class GIPCBufferedOpusEncoder(OpusEncoder):
-    FIN = 1
-
-    def __init__(self, source, *args, **kwargs):
-        import gipc
-
-        self.source = source
-        self.parent_pipe, self.child_pipe = gipc.pipe(duplex=True)
-        self.frames = Queue(kwargs.pop('queue_size', 4096))
-        super(GIPCBufferedOpusEncoder, self).__init__(*args, **kwargs)
-
-        gipc.start_process(target=self._encoder_loop, args=(self.child_pipe, (args, kwargs)))
-
-        gevent.spawn(self._writer)
-        gevent.spawn(self._reader)
-
-    def _reader(self):
-        while True:
-            data = self.parent_pipe.get()
-            if data == self.FIN:
-                return
-
-            self.frames.put(data)
-        self.parent_pipe = None
-
-    def _writer(self):
-        while self.data:
-            raw = self.source.read(self.frame_size)
-            if len(raw) < self.frame_size:
-                break
-
-            self.parent_pipe.put(raw)
-            gevent.idle()
-
-        self.parent_pipe.put(self.FIN)
-
-    def have_frame(self):
-        return self.parent_pipe
-
-    def next_frame(self):
-        return self.frames.get()
-
-    @classmethod
-    def _encoder_loop(cls, pipe, (args, kwargs)):
-        encoder = OpusEncoder(*args, **kwargs)
-
-        while True:
-            data = pipe.get()
-            if data == cls.FIN:
-                pipe.put(cls.FIN)
-                return
-
-            pipe.put(encoder.encode(data, encoder.samples_per_frame))
-
-
-class DCADOpusEncoder(OpusEncoder):
-    def __init__(self, source, *args, **kwargs):
-        self.source = source
-        self.command = kwargs.pop('command', 'dcad')
-        super(DCADOpusEncoder, self).__init__(*args, **kwargs)
-        self._proc = None
-        self.header_size = struct.calcsize('<h')
-
-    @property
-    def proc(self):
-        if not self._proc:
-            self._proc = subprocess.Popen([
-                self.command,
-                '--channels', str(self.channels),
-                '--rate', str(self.sampling_rate),
-                '--size', str(self.samples_per_frame),
-                '--bitrate', '128',
-                '--fec',
-                '--packet-loss-percent', '30',
-                '--input', 'pipe:0',
-                '--output', 'pipe:1',
-            ], stdin=self.source.stdout, stdout=subprocess.PIPE)
-        return self._proc
-
-    def have_frame(self):
-        return bool(self.proc)
-
-    def next_frame(self):
-        header = self.proc.stdout.read(self.header_size)
-        if len(header) < self.header_size:
-            self._proc = None
-            return
-
-        size = struct.unpack('<h', header)[0]
-
-        data = self.proc.stdout.read(size)
-        if len(data) == 0:
-            self._proc = None
-            return
-
-        return data
